feat(junit): Improve argument converters
