+++ utils/config-utils/src/test/java/datadog/trace/api/ConfigSettingTest.java
@@ -3,8 +3,10 @@ package datadog.trace.api;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertNotEquals;
 
-import datadog.trace.junit.utils.tabletest.BoxedValueConverter;
 import datadog.trace.junit.utils.tabletest.ConfigValueConverter;
+import org.junit.jupiter.api.extension.ParameterContext;
+import org.junit.jupiter.params.converter.ArgumentConversionException;
+import org.junit.jupiter.params.converter.ArgumentConverter;
 import org.junit.jupiter.params.converter.ConvertWith;
 import org.tabletest.junit.TableTest;
 
@@ -79,5 +81,59 @@ public class ConfigSettingTest {
   void convertIterableMapAndBitSetToString(
       @ConvertWith(ConfigValueConverter.class) Object value, String rendered) {
     assertEquals(rendered, ConfigSetting.of("key", value, ConfigOrigin.DEFAULT).stringValue());
+  }
+
+  /**
+   * Converts a String cell value to the most specific boxed Java primitive type. Use with
+   * {@code @ConvertWith(BoxedValueConverter.class)} on {@code Object}-typed parameters when the test
+   * needs actual typed values (e.g. {@code Float} not {@code String "2.33f"}).
+   *
+   * <p>Conversion rules:
+   *
+   * <ul>
+   *   <li>blank/null -> null
+   *   <li>{@code "true"}/{@code "false"} -> {@link Boolean}
+   *   <li>ends with {@code "f"} -> {@link Float}
+   *   <li>contains {@code "."} -> {@link Double}
+   *   <li>parseable as integer -> {@link Integer}
+   *   <li>otherwise -> {@link String}
+   * </ul>
+   */
+  static class BoxedValueConverter implements ArgumentConverter {
+    @Override
+    public Object convert(Object source, ParameterContext context)
+        throws ArgumentConversionException {
+      if (source == null) {
+        return null;
+      }
+
+      String s = source.toString();
+      switch (s) {
+        case "":
+          return null;
+        case "true":
+          return Boolean.TRUE;
+        case "false":
+          return Boolean.FALSE;
+      }
+      if (s.endsWith("f")) {
+        try {
+          return Float.parseFloat(s.substring(0, s.length() - 1));
+        } catch (NumberFormatException ignored) {
+        }
+      }
+
+      if (s.contains(".")) {
+        try {
+          return Double.parseDouble(s);
+        } catch (NumberFormatException ignored) {
+        }
+      }
+      try {
+        return Integer.parseInt(s);
+      } catch (NumberFormatException ignored) {
+      }
+      return s;
+    }
   }
 }
